Code changes to support boostrap 4
