+++ lib/rspec_html_reporter.rb
@@ -1,168 +1,13 @@
 # frozen_string_literal: true
 
-require 'rspec/core/formatters/base_formatter'
 require 'active_support'
 require 'active_support/core_ext/numeric'
 require 'active_support/inflector'
 require 'fileutils'
-require 'rouge'
 require 'erb'
-require 'rbconfig'
+require 'rspec_html_reporter/example'
 
 I18n.enforce_available_locales = false
-
-class Oopsy
-  attr_reader :klass, :message, :backtrace, :highlighted_source, :explanation, :backtrace_message
-
-  def initialize(example, file_path)
-    @example = example
-    @exception = @example.exception
-    @file_path = file_path
-    unless @exception.nil?
-      @klass = @exception.class
-      @message = @exception.message.encode('utf-8')
-      @backtrace = @exception.backtrace
-      @backtrace_message = formatted_backtrace(@example, @exception)
-      @highlighted_source = process_source
-      @explanation = process_message
-    end
-  end
-
-  private
-
-  def os
-    @os ||= begin
-      host_os = RbConfig::CONFIG['host_os']
-      case host_os
-      when /mswin|msys|mingw|cygwin|bccwin|wince|emc/
-        :windows
-      when /darwin|mac os/
-        :macosx
-      when /linux/
-        :linux
-      when /solaris|bsd/
-        :unix
-      else
-        raise StandardError, "unknown os: #{host_os.inspect}"
-      end
-    end
-  end
-
-  def formatted_backtrace(example, exception)
-    # To avoid an error in format_backtrace. RSpec's version below v3.5 will throw exception.
-    return [] unless example
-
-    formatter = RSpec.configuration.backtrace_formatter
-    formatter.format_backtrace(exception.backtrace, example.metadata)
-  end
-
-  def process_source
-    return '' if @backtrace_message.empty?
-
-    data = @backtrace_message.first.split(':')
-    unless data.empty?
-      if os == :windows
-        file_path = "#{data[0]}:#{data[1]}"
-        line_number = data[2].to_i
-      else
-        file_path = data.first
-        line_number = data[1].to_i
-      end
-      lines = File.readlines(file_path)
-      start_line = line_number - 2
-      end_line = line_number + 3
-      source = lines[start_line..end_line].join('').sub(lines[line_number - 1].chomp,
-                                                        "--->#{lines[line_number - 1].chomp}")
-
-      formatter = Rouge::Formatters::HTMLLegacy.new(css_class: 'highlight', line_numbers: true,
-                                                    start_line: start_line + 1)
-      lexer = Rouge::Lexers::Ruby.new
-      formatter.format(lexer.lex(source.encode('utf-8')))
-    end
-  end
-
-  def process_message
-    formatter = Rouge::Formatters::HTMLLegacy.new(css_class: 'highlight')
-    lexer = Rouge::Lexers::Ruby.new
-    formatter.format(lexer.lex(@message))
-  end
-end
-
-class Example
-  def self.load_spec_comments!(examples)
-    examples.group_by(&:file_path).each do |file_path, file_examples|
-      lines = File.readlines(file_path)
-
-      file_examples.zip(file_examples.rotate).each do |ex, next_ex|
-        lexically_next = next_ex &&
-                         next_ex.file_path == ex.file_path &&
-                         next_ex.metadata[:line_number] > ex.metadata[:line_number]
-        start_line_idx = ex.metadata[:line_number] - 1
-        next_start_idx = (lexically_next ? next_ex.metadata[:line_number] : lines.size) - 1
-        spec_lines = lines[start_line_idx...next_start_idx].select { |l| l.match(/#->/) }
-        ex.set_spec(spec_lines.join) unless spec_lines.empty?
-      end
-    end
-  end
-
-  attr_reader :example_group, :description, :full_description, :run_time, :duration, :status, :exception, :file_path,
-              :metadata, :spec, :screenshots, :screenrecord, :failed_screenshot
-
-  def initialize(example)
-    @example_group = example.example_group.to_s
-    @description = example.description
-    @full_description = example.full_description
-    @execution_result = example.execution_result
-    @run_time = (@execution_result.run_time).round(5)
-    @duration = @execution_result.run_time.to_fs(:rounded, precision: 5)
-    @status = @execution_result.status.to_s
-    @metadata = example.metadata
-    @file_path = @metadata[:file_path]
-    @exception = Oopsy.new(example, @file_path)
-    @spec = nil
-    @screenshots = @metadata[:screenshots]
-    @screenrecord = @metadata[:screenrecord]
-    @failed_screenshot = @metadata[:failed_screenshot]
-  end
-
-  def example_title
-    title_arr = @example_group.to_s.split('::') - %w[RSpec ExampleGroups]
-    title_arr.push @description
-
-    title_arr.join(' → ')
-  end
-
-  def has_exception?
-    !@exception.klass.nil?
-  end
-
-  def has_spec?
-    !@spec.nil?
-  end
-
-  def has_screenshots?
-    !@screenshots.nil? && !@screenshots.empty?
-  end
-
-  def has_screenrecord?
-    !@screenrecord.nil?
-  end
-
-  def has_failed_screenshot?
-    !@failed_screenshot.nil?
-  end
-
-  def set_spec(spec_text)
-    formatter = Rouge::Formatters::HTMLLegacy.new(css_class: 'highlight')
-    lexer = Rouge::Lexers::Gherkin.new
-    @spec = formatter.format(lexer.lex(spec_text.gsub('#->', '')))
-  end
-
-  def klass(prefix = 'label-')
-    class_map = { passed: "#{prefix}success", failed: "#{prefix}danger", pending: "#{prefix}warning" }
-    class_map[@status.to_sym]
-  end
-end
 
 class RspecHtmlReporter < RSpec::Core::Formatters::BaseFormatter
   DEFAULT_REPORT_PATH = File.join(Bundler.root, 'reports', Time.now.strftime('%Y%m%d-%H%M%S'))
@@ -203,6 +48,8 @@ class RspecHtmlReporter < RSpec::Core::Formatters::BaseFormatter
   end
 
   def example_passed(notification)
+    # require 'byebug'; byebug
+
     @group_example_success_count += 1
     @examples << Example.new(notification.example)
   end
